Create remaining.ino
In this example, the Every object is created with an interval of 5 seconds, and a callback function for the first event is set. The update() method is called in the loop() function to check if the interval has passed. Inside the loop function, the remaining() method is used to check if the time remaining before the next trigger is less than or equal to half of the interval, if true, the secondEvent function is called.