Create complexExample.ino
In this example, 3 instances of the Every class are created: one for a 1-second interval, one for a 5-second interval, and one for a 10-second interval. The setup() function sets the callback functions for each event and sets the repeat option for the third event to false so it will happen once.

The loop() function calls the update() method for all the timers. The first event (firstEvent) triggers every second and blinks the LED, the second event (secondEvent) triggers every five seconds and prints a message on the serial monitor, the third event (thirdEvent) triggers every ten seconds, and change the interval of the first event to blink every 2 seconds, the fourth event (fourthEvent) triggers right after the third event and reset the first event to blink every 1 second.